+++ crates/observability/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+name = "observability"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+http = { version = "1.2.0", optional = true }
+opentelemetry = { version = "0.27.1", optional = true }
+opentelemetry-appender-tracing = { version = "0.27.0", optional = true }
+opentelemetry-http = { version = "0.27.0", optional = true }
+opentelemetry-otlp = { version = "0.27.0", features = ["logs", "metrics"], optional = true }
+opentelemetry-semantic-conventions = { version = "0.27.0", optional = true }
+opentelemetry_sdk = { version = "0.27.1", features = ["rt-tokio"], optional = true }
+tokio = { version = "1.43.0", default-features = false, features = ["signal"], optional = true }
+tonic = { version = "0.12.3", default-features = false, optional = true }
+tower-http = { version = "0.6.2", features = ["trace"], optional = true }
+tracing = { version = "0.1.41", optional = true }
+tracing-opentelemetry = { version = "0.28.0", optional = true }
+tracing-subscriber = { version = "0.3.19", features = ["env-filter"], optional = true }
+
+[features]
+default = []
+server = [
+    "dep:http",
+    "dep:opentelemetry",
+    "dep:opentelemetry-http",
+    "dep:tokio",
+    "dep:tonic",
+    "dep:tower-http",
+    "dep:tracing",
+    "dep:tracing-opentelemetry",
+    "opentelemetry-semantic-conventions/semconv_experimental"
+]
+provider = [
+    "dep:opentelemetry",
+    "dep:opentelemetry-appender-tracing",
+    "dep:opentelemetry-otlp",
+    "dep:opentelemetry-semantic-conventions",
+    "dep:opentelemetry_sdk",
+    "dep:tracing",
+    "dep:tracing-opentelemetry",
+    "dep:tracing-subscriber",
+]
+
+[lints.clippy]
+pedantic = "warn"
