+++ services/tenant/Cargo.toml
@@ -8,16 +8,9 @@ adapter = { version = "0.1.0", path = "internal/adapter" }
 app = { version = "0.1.0", path = "internal/app" }
 aws-config = "1.5.13"
 driver = { version = "0.1.0", path = "internal/driver" }
-kernel = { version = "0.1.0", path = "internal/kernel" }
-opentelemetry = "0.27.1"
-opentelemetry-appender-tracing = "0.27.0"
-opentelemetry-otlp = { version = "0.27.0", features = ["logs", "metrics"] }
-opentelemetry-semantic-conventions = "0.27.0"
-opentelemetry_sdk = { version = "0.27.1", features = ["rt-tokio"] }
+observability = { version = "0.1.0", path = "../../crates/observability", features = ["provider"] }
 tokio = { version = "1.41.0", default-features = false, features = ["macros", "rt-multi-thread"] }
 tracing = "0.1.41"
-tracing-opentelemetry = "0.28.0"
-tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 
 [dev-dependencies]
 aws-sdk-dynamodb = "1.59.0"
