+++ services/tenant/internal/driver/Cargo.toml
@@ -5,19 +5,14 @@ edition = "2021"
 
 [dependencies]
 app = { version = "0.1.0", path = "../app" }
-http = "1.2.0"
-opentelemetry = "0.27.1"
-opentelemetry-http = "0.27.0"
-opentelemetry-semantic-conventions = { version = "0.27.0", features = ["semconv_experimental"] }
+observability = { version = "0.1.0", path = "../../../../crates/observability", features = ["server"] }
 proto = { version = "0.1.0", path = "../../../../proto/generate/rust", default-features = false, features = ["tenant-v1"] }
-tokio = { version = "1.41.0", default-features = false, features = ["signal"] }
 tonic = "0.12.3"
 tonic-health = "0.12.3"
 tonic-reflection = "0.12.3"
 tonic-types = "0.12.3"
-tower-http = { version = "0.6.2", features = ["catch-panic", "trace"] }
+tower-http = { version = "0.6.2", features = ["catch-panic"] }
 tracing = "0.1.40"
-tracing-opentelemetry = "0.28.0"
 
 [lints]
 workspace = true
