+++ services/tenant/internal/driver/src/server.rs
@@ -2,7 +2,6 @@ use std::fmt::Debug;
 use std::net::SocketAddr;
 
 use app::{CommandUseCaseExt, QueryUseCaseExt};
-use opentelemetry::trace::TraceContextExt;
 use proto::tenant::v1::tenant_service_server::{TenantService, TenantServiceServer};
 use proto::tenant::v1::{
     AddItemsRequest, AddItemsResponse, CreateRequest, CreateResponse, ListItemsRequest,
@@ -12,7 +11,6 @@ use proto::tenant::v1::{
 use tonic::{Code, Request, Response, Status};
 use tonic_types::{ErrorDetails, StatusExt};
 use tracing::instrument;
-use tracing_opentelemetry::OpenTelemetrySpanExt;
 
 type Error = Box<dyn std::error::Error + Send + Sync + 'static>;
 
@@ -219,7 +217,6 @@ where
         use std::time::Duration;
 
         use tower_http::catch_panic::CatchPanicLayer;
-        use tower_http::trace::TraceLayer;
 
         let tenant_service = TenantServiceServer::new(self.service);
         let reflection_service = tonic_reflection::server::Builder::configure()
@@ -241,168 +238,12 @@ where
                     Status::unknown(err).into_http()
                 },
             ))
-            .layer(
-                TraceLayer::new_for_grpc()
-                    .make_span_with(MakeSpan)
-                    .on_request(OnRequest)
-                    .on_response(OnResponse)
-                    .on_failure(OnFailure),
-            )
+            .layer(observability::server::grpc_trace_layer())
             .add_service(tenant_service)
             .add_service(reflection_service)
             .add_service(health_service)
-            .serve_with_shutdown(addr, shutdown_signal())
+            .serve_with_shutdown(addr, observability::server::shutdown())
             .await?;
         Ok(())
-    }
-}
-
-/// サーバーを安全に終了するための仕組み(Graceful shutdown)
-async fn shutdown_signal() {
-    use tokio::signal;
-
-    let ctrl_c = async {
-        signal::ctrl_c()
-            .await
-            .unwrap_or_else(|e| panic!("failed to install Ctrl+C handler: {e}"));
-    };
-    let terminate = async {
-        signal::unix::signal(signal::unix::SignalKind::terminate())
-            .unwrap_or_else(|e| panic!("failed to install signal handler: {e}"))
-            .recv()
-            .await;
-    };
-    tokio::select! {
-        () = ctrl_c => tracing::trace!("receive ctrl_c signal"),
-        () = terminate => tracing::trace!("receive terminate"),
-    }
-}
-
-#[derive(Clone)]
-struct MakeSpan;
-
-impl<B> tower_http::trace::MakeSpan<B> for MakeSpan {
-    fn make_span(&mut self, req: &http::Request<B>) -> tracing::Span {
-        use tracing_opentelemetry::OpenTelemetrySpanExt as _;
-
-        if req.uri().path().contains("ServerReflectionInfo") {
-            // NOTE: リフレクションのトレースは出力しない
-            return tracing::Span::none();
-        }
-        let span = tracing::info_span!("", otel.name = %req.uri().path()[1..]);
-
-        let ctx = opentelemetry::global::get_text_map_propagator(|p| {
-            p.extract(&opentelemetry_http::HeaderExtractor(req.headers()))
-        });
-        if ctx.span().span_context().is_valid() {
-            span.set_parent(ctx);
-        }
-
-        span
-    }
-}
-
-#[derive(Clone)]
-struct OnRequest;
-
-impl<B> tower_http::trace::OnRequest<B> for OnRequest {
-    fn on_request(&mut self, req: &http::Request<B>, span: &tracing::Span) {
-        use opentelemetry_semantic_conventions::trace::{
-            NETWORK_PEER_ADDRESS, NETWORK_PEER_PORT, RPC_GRPC_REQUEST_METADATA, RPC_METHOD,
-            RPC_SERVICE, RPC_SYSTEM,
-        };
-        use tracing_opentelemetry::OpenTelemetrySpanExt as _;
-
-        span.set_attribute(RPC_SYSTEM, "grpc");
-        let uri = req.uri().clone();
-        {
-            let v: Vec<_> = uri.path()[1..].split('/').collect();
-            if v.len() == 2 {
-                if let Some(service) = v.first() {
-                    span.set_attribute(RPC_SERVICE, (*service).to_string());
-                }
-                if let Some(method) = v.get(1) {
-                    span.set_attribute(RPC_METHOD, (*method).to_string());
-                }
-            }
-        }
-        if let Some(host) = uri.host() {
-            span.set_attribute(NETWORK_PEER_ADDRESS, (*host).to_string());
-        }
-        if let Some(port) = uri.port() {
-            span.set_attribute(NETWORK_PEER_PORT, port.to_string());
-        }
-        for (key, value) in req.headers().clone() {
-            let (key, value) = match (key, value.to_str()) {
-                (Some(key), Ok(value)) => (key, value.to_string()),
-                _ => continue,
-            };
-            span.set_attribute(format!("{RPC_GRPC_REQUEST_METADATA}.{key}"), value);
-        }
-    }
-}
-
-#[derive(Clone)]
-struct OnResponse;
-
-impl<B> tower_http::trace::OnResponse<B> for OnResponse {
-    fn on_response(self, res: &http::Response<B>, _: std::time::Duration, span: &tracing::Span) {
-        use opentelemetry_semantic_conventions::trace::{
-            RPC_GRPC_RESPONSE_METADATA, RPC_GRPC_STATUS_CODE,
-        };
-        use tracing_opentelemetry::OpenTelemetrySpanExt as _;
-
-        let headers = res.headers().clone();
-        for (key, value) in headers {
-            let (key, value) = match (key, value.to_str()) {
-                (Some(key), Ok(value)) => (key, value.to_string()),
-                _ => continue,
-            };
-            span.set_attribute(format!("{RPC_GRPC_RESPONSE_METADATA}.{key}"), value.clone());
-        }
-        match tonic::Status::from_header_map(res.headers()) {
-            Some(status) if status.code() != tonic::Code::Ok => {
-                span.set_attribute(RPC_GRPC_STATUS_CODE, i64::from(i32::from(status.code())));
-            }
-            _ => span.set_attribute(RPC_GRPC_STATUS_CODE, 0),
-        }
-    }
-}
-
-#[derive(Clone)]
-struct OnFailure;
-
-impl<F: std::any::Any> tower_http::trace::OnFailure<F> for OnFailure {
-    fn on_failure(
-        &mut self,
-        failure_classification: F,
-        _: std::time::Duration,
-        span: &tracing::Span,
-    ) {
-        if let Some(value) = (&failure_classification as &dyn std::any::Any)
-            .downcast_ref::<tower_http::classify::GrpcFailureClass>()
-        {
-            match value {
-                tower_http::classify::GrpcFailureClass::Code(code) => {
-                    match tonic::Code::from_i32(code.get()) {
-                        Code::Ok
-                        | Code::Cancelled
-                        | Code::InvalidArgument
-                        | Code::NotFound
-                        | Code::AlreadyExists
-                        | Code::PermissionDenied
-                        | Code::ResourceExhausted
-                        | Code::FailedPrecondition
-                        | Code::Aborted
-                        | Code::OutOfRange
-                        | Code::Unauthenticated => {
-                            span.set_status(opentelemetry::trace::Status::Unset);
-                        }
-                        _ => (),
-                    }
-                }
-                tower_http::classify::GrpcFailureClass::Error(_) => (),
-            }
-        }
     }
 }
