+++ services/tenant/src/main.rs
@@ -3,13 +3,10 @@ use app::{CommandUseCase, QueryUseCase};
 use aws_config::BehaviorVersion;
 use driver::server::{Server, Service};
 
-use self::observability::instrument;
-
-mod observability;
-
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
-    let shutdown_providers = instrument()?;
+    let shutdown_providers =
+        observability::provider::init_providers(env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"))?;
     let addr = format!(
         "[::1]:{}",
         std::env::var("PORT").map_err(|e| format!("PORT must be set: {e:?}"))?
